fixed at least, rain is buggy,  rain is sometimes comming nil from json, description was fixed, models are quaty correct now

+++ WeatherForecast-DataLayer/Models/SubModels/WeatherDetail.swift
@@ -12,7 +12,7 @@ import Foundation
 struct WeatherDetail: Codable {
     let temp, tempMin, tempMax, pressure: Double
     let seaLevel, grndLevel: Double
-    let humidity, tempKf: Int
+    let humidity, tempKf: Double
     
     enum CodingKeys: String, CodingKey {
         case temp
